==> NLP/2-corpus-preprocessing.ipynb <==
# Módulo 2: Preprocesado del Texto

En esta etapa, nos dedicaremos a preprocesar el texto. Haremos todas las modificaciones pertinentes para que pueda ser estudiado más a profundidad por modelos de Machine Learning. Ahora crearemos una función que se encargará de realizar este preprocesado.

# En primer lugar, hacemos los imports necesarios
import re
import string
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer

# Descargamos los paquetes necesarios
nltk.download('punkt')
nltk.download('stopwords')
nltk.download('wordnet')
nltk.download('omw-1.4')

# Creamos la función que se encargará de realizar el preprocesado del texto
def preprocess(text, listaStopWords):

  # Paso 1: Convertimos todo el texto a minúscula
  text = text.lower()

  # Paso 2: Eliminar signos de puntuación y caracteres no alfabéticos
  text = re.sub('[%s]' % re.escape(string.punctuation), '', text)
  text = re.sub(r'[^a-zA-Z]', ' ', text)

  # Paso 3: Tokenización de palabras
  tokens = nltk.word_tokenize(text)

  # Paso 4: Lematización
  lemmatizer = WordNetLemmatizer()
  tokens = [lemmatizer.lemmatize(token) for token in tokens]

  # Eliminación de las stopwords
  sw = stopwords.words('english')
  sw.extend(listaStopWords)
  conjuntoSW = set(sw)
  tokens = [token for token in tokens if token not in conjuntoSW]

  # Unión de los tokens en un texto
  textoProcesado = ' '.join(tokens)

  # Devuelta del texto preprocesado
  return textoProcesado

Una vez teniendo nuestra función, encargada de preprocesar el texto, crearemos una función que se encargará de aplicarla a cada uno de nuestros reviews.

# Creamos la función
def ReviewsAplication(reviews, listaStopWords):

  # Inicializamos una lista vacía que contendrá nuestros reviews procesados
  reviewsProcesados = []

  # Creamos el bucle "for" que se encargará de aplicar nuestra función anterior a cada review y añadirlo a la lista
  for texto in reviews:
    texto = preprocess(texto, listaStopWords)
    reviewsProcesados.append(texto)

  # Devuelta la lista de reviews procesados
  return reviewsProcesados

Ya teniendo estas funciones, podemos importar el DataFrame que elaboramos en la parte de Descarga y exploración del corpus para que pueda ser debidamente preparado para la parte de Machine Learning.

# Hacemos los imports necesarios
import pandas as pd

# Importamos el csv con el dfcorpus
dfcorpus = pd.read_csv('./MusicalInstruments.csv')

# Nos aseguramos que el importe se ha realizado correctamente
dfcorpus.head()

Viendo que la importación ha ocurrido sin ningún problema, podemos aplicar el preprocesamiento a nuestro corpus con las funciones creadas anteriormente.

# Primero, seleccionamos nuestras stopwords basándonos en los resultados dados por el wordcloud
listaStopWords = ['string', 'time', 'guitar', 'play', 'one', 'need', 'work', 'sound', 'pedal', 'got', 'will', 'price', 'mic']

# Creamos una nueva variable compuesta de los reviews del corpus ya procesados
corpusProcesado = ReviewsAplication(dfcorpus['reviewText'], listaStopWords)

Ahora crearemos un nuevo dfcorpus donde el campo reviewText contendrá los reviews procesados.

# Creamos una copia del dfcorpus antes del preprocesamiento
dfcorpusProcesado = dfcorpus.copy()

# Sustituimos la información del campo reviewText por la misma información pero ya preprocesada
dfcorpusProcesado['reviewText'] = corpusProcesado

# Mostramos una parte del DataFrame para asegurarnos que los cambios han ocurrido correctamente
dfcorpusProcesado.head()

### Conclusiones del Preprocesado.

Ahora tenemos un DataFrame con informaciones más relevantes para nuestro Supervised Sentiment Analysis. Con este corpus libre de signos de puntuación, caracteres no alfabéticos, stopwords y con cada palabra totalmente en minúscula, podemos sentirnos cómodos para pasar a la etapa de modelado.

# Guardo el DataFrame en un archivo csv para poder utilizarlo en partes posteriores
dfcorpusProcesado.to_csv('MusicalInstrumentsPreprocesado.csv', index=False)
