==> NLP/3-corpus-modeling.ipynb <==
# Módulo 3: Entrenamiento y Testeo de un modelo de análisis de Sentimiento

En esta etapa nos dedicaremos a entrenar y probar un modelo de anaálisis de sentimiento. Aquí entrenaremos dos modelos distintos de Machine Learning y veremos, mediante diferentes métricas, cual de ellos arroja un mejor resultado.

# En primer lugar, hacemos los imports necesarios
import pandas as pd

# Luego importamos el dfcorpus preprocesado de la etapa anterior
dfcorpus = pd.read_csv('./MusicalInstrumentsPreprocesado.csv')

# Nos aseguramos que el importe se ha realizado correctamente
dfcorpus.head()

Ahora aseguremonos por última vez que nuestro DataFrame no presenta datos vacíos.

dfcorpus.info()

Podemos observar que tenemos algunos valores nulos producto del preprocesado. Vamos a eliminar estos registros antes de realizar el split. Podemos hacerlo sin mayor problema porque son pocos en comparación a la cantidad de información que poseemos.

# Eliminamos los registros donde el campo reviewText se encuentra vacío
dfcorpus.dropna(inplace=True)

# Nos aseguramos que el proceso ha ocurrido correctamente
dfcorpus.info()

Ahora realizaremos la separación en conjunto de entrenamiento y conjunto de prueba.

# Hacemos los imports necesarios para la separación
from sklearn.model_selection import train_test_split

# Hacemos la separación en entrenamiento y prueba para la variable predictora y la variable a predecir (test_size = 10%)
xtrain, xtest, ytrain, ytest = train_test_split(dfcorpus['reviewText'], dfcorpus['overall'], test_size=0.1, random_state=42, shuffle=True)

# Ahora veremos las dimensiones de nuestras variables
print(f'Shape xtrain: {xtrain.shape}')
print(f'Shape ytrain: {ytrain.shape}')
print(f'Shape xtest: {xtest.shape}')
print(f'Shape ytest: {ytest.shape}')

A continuación haremos una extracción de características mediante la implementación de codificación basado en Bolsa de Palabras (BoW). Para eso utilizaremos la clase de Scikit-Learn CountVectorizer.

# En primer lugar, hacemos los imports necesarios
from sklearn.feature_extraction.text import CountVectorizer

# Ahora creamos el objeto
cv = CountVectorizer(max_features=2000, ngram_range=(1,2), max_df=0.95, min_df=5)

# Ajustamos y transformamos a nuestra variable predictora de entrenamiento
xtrain_cv = cv.fit_transform(xtrain)

# Utilizamos el ajuste realizado con las variables predictoras de entrenamiento para transformar las variables predictoras de prueba
xtest_cv = cv.transform(xtest)

En este momento aplicaremos un modelo de Regresión Logística y veremos que tan buenos resultados nos trae.

# En primer lugar, hacemos los imports necesarios
from sklearn.model_selection import GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score

# Ahora creamos la lista de posibles hiperparámetros con el fin de decidir el óptimo
possibleMaxIter = [500, 1000, 2000, 5000]

# Luego construimos y ajustamos el GridSearchCV
param_grid = {'max_iter' : possibleMaxIter}
grid = GridSearchCV(LogisticRegression(random_state=42), param_grid=param_grid, cv=10, verbose=2)
grid.fit(xtrain_cv, ytrain)

# Posteriormente inicializamos los hiperparámetros óptimos
maxIterOptimo = grid.best_params_['max_iter']

# Ahora inicializamos el mismo modelo nuevamente, pero esta vez con los parámetros óptimos
lr = LogisticRegression(random_state=42, max_iter=maxIterOptimo)
lr.fit(xtrain_cv, ytrain)

# -------------------------------- HACEMOS LAS PREDICCIONES --------------------------------

# Hacemos las predicciones a ytrain y ytest
lr_ytrain_pred = lr.predict(xtrain_cv)
lr_ytest_pred = lr.predict(xtest_cv)

# -------------------------------- CALCULAMOS LAS MÉTRICAS --------------------------------

# Calculamos el accuracy
print('Accuracy')
print(f'Train: {accuracy_score(ytrain, lr_ytrain_pred)}')
print(f'Test: {accuracy_score(ytest, lr_ytest_pred)} \n')

# Calculamos la precision
print('Precision')
print(f'Train: {precision_score(ytrain, lr_ytrain_pred)}')
print(f'Test: {precision_score(ytest, lr_ytest_pred)} \n')

# Calculamos el recall
print('Recall')
print(f'Train: {recall_score(ytrain, lr_ytrain_pred)}')
print(f'Test: {recall_score(ytest, lr_ytest_pred)} \n')

# Calculamos el f1_score
print('F1 Score')
print(f'Train: {f1_score(ytrain, lr_ytrain_pred)}')
print(f'Test: {f1_score(ytest, lr_ytest_pred)} \n')

Podemos observar que obtenemos resultados bastantes decentes y sin overfitting aplicando una Regresión Logística sencilla. Esto puede deberse a que tenemos suficientes datos para entrenar cómodamente al modelo y al hecho de que se hizo un buen preprocesado de texto...

Ahora veremos que tan buenos resultados podemos obtrener con un modelo de Clasificación Binaria con Bosques Aleatorios.

# En primer lugar, hacemos los imports necesarios
from sklearn.model_selection import GridSearchCV
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score

# Ahora creamos la lista de posibles hiperparámetros con el fin de decidir el óptimo
possibleMaxDepth = range(1,6)

# Luego construimos y ajustamos el GridSearchCV
param_grid = {'max_depth' : possibleMaxDepth}
grid = GridSearchCV(RandomForestClassifier(random_state=42), param_grid=param_grid, cv=10, verbose=2)
grid.fit(xtrain_cv, ytrain)

# Posteriormente inicializamos los hiperparámetros óptimos
maxDepthOptimo = grid.best_params_['max_depth']

# Ahora inicializamos el mismo modelo nuevamente, pero esta vez con los parámetros óptimos
rfc = RandomForestClassifier(random_state=42, max_depth=maxDepthOptimo)
rfc.fit(xtrain_cv, ytrain)

# -------------------------------- HACEMOS LAS PREDICCIONES --------------------------------

# Hacemos las predicciones a ytrain y ytest
rfc_ytrain_pred = rfc.predict(xtrain_cv)
rfc_ytest_pred = rfc.predict(xtest_cv)

# -------------------------------- CALCULAMOS LAS MÉTRICAS --------------------------------

# Calculamos el accuracy
print('Accuracy')
print(f'Train: {accuracy_score(ytrain, rfc_ytrain_pred)}')
print(f'Test: {accuracy_score(ytest, rfc_ytest_pred)} \n')

# Calculamos la precision
print('Precision')
print(f'Train: {precision_score(ytrain, rfc_ytrain_pred)}')
print(f'Test: {precision_score(ytest, rfc_ytest_pred)} \n')

# Calculamos el recall
print('Recall')
print(f'Train: {recall_score(ytrain, rfc_ytrain_pred)}')
print(f'Test: {recall_score(ytest, rfc_ytest_pred)} \n')

# Calculamos el f1_score
print('F1 Score')
print(f'Train: {f1_score(ytrain, rfc_ytrain_pred)}')
print(f'Test: {f1_score(ytest, rfc_ytest_pred)} \n')

Podemos observar que obtenemos resultados bastantes decentes y sin overfitting aplicando una Clasificación con Bosques Aleatorios sencilla. Esto puede deberse a que tenemos suficientes datos para entrenar cómodamente al modelo y al hecho de que se hizo un buen preprocesado de texto...

### Conclusiones del Modelado

En esta etapa probamos dos modelos de Machine Learning, siendo estos Regresión Logística y Clasificación con Bosques Aleatorios. Si bien es cierto que ambos modelos arrojaron resultados bastante decentes, pudimos notar que la Regresión Logística arrojó resultados ligeramente mejores que la Clasificación con Bosques Aleatorios. Rescataremos las métricas calculadas y estudiaremos sus resultados en la etapa posterior.

# Creamos un diccionario con las métricas calculadas de los modelos
metrics_dict = {'Model':['LogisticRegression','LogisticRegression','RandomForestClassifier','RandomForestClassifier'],
                'Category':['Train','Test','Train','Test'],
                'Accuracy':[accuracy_score(ytrain, lr_ytrain_pred), accuracy_score(ytest, lr_ytest_pred), accuracy_score(ytrain, rfc_ytrain_pred), accuracy_score(ytest, rfc_ytest_pred)],
                'Precision':[precision_score(ytrain, lr_ytrain_pred), precision_score(ytest, lr_ytest_pred), precision_score(ytrain, rfc_ytrain_pred), precision_score(ytest, rfc_ytest_pred)],
                'Recall':[recall_score(ytrain, lr_ytrain_pred), recall_score(ytest, lr_ytest_pred), recall_score(ytrain, rfc_ytrain_pred), recall_score(ytest, rfc_ytest_pred)],
                'F1 Score':[f1_score(ytrain, lr_ytrain_pred), f1_score(ytest, lr_ytest_pred), f1_score(ytrain, rfc_ytrain_pred), f1_score(ytest, rfc_ytest_pred)]}

# Ahora creamos un DataFrame a partir del diccionario creado anteriormente
dfmetrics = pd.DataFrame(metrics_dict)

# Ahora nos aseguramos que todo se generó de la manera esperada
dfmetrics.head()

# Guardo el DataFrame en un archivo csv para poder utilizarlo en partes posteriores
dfmetrics.to_csv('MusicalInstrumentsMetrics.csv', index=False)
