==> NLP/4-corpus-conclusions.ipynb <==
# Módulo 4: Reporte de métricas y conclusiones

En este momento, se rescatarán y se estudiarán los resultados obtenidos por los modelos. Para eso importaremos el DataFrame creado en la parte de modelado con el reporte de las métricas calculadas.

# En primer lugar, hacemos los imports necesarios
import pandas as pd

# Luego importamos el Reporte de Métricas de la etapa anterior
dfmetrics = pd.read_csv('./MusicalInstrumentsMetrics.csv')

# Agrupamos por modelo
dfmetrics.groupby(['Model','Category']).mean()

### Conclusiones Finales

Este proyecto de Supervised Sentiment Analysis se compuso de 4 partes: Descarga y exploracion del corpus, Preprocesamiento del corpus, Modelado del corpus y Reporte de métricas. En cada una de las partes se ejecutaron procesos importantes y se descubrieron cosas relevantes. Las observaciones más marcadas serán descritas a continuación:

- Descarga y exploración del corpus:
En esta etapa estudiamos las informaciones del corpus. Notamos que estaba bastante desbalanceado en el sentido de que había demasiadas reviews de calificación 5 en comparación a las demás calificaciones. Por eso tomamos la decisión de tratarlo como un problema de clasificación binaria donde cualquier calificación inferior a 5 sería considerada como calificación negativa. Aún así el corpus estaba desbalanceado, sin embargo, ya era mucho más cómodo de trabajar.

- Preprocesamiento del corpus:
En esta etapa logramos reestructurar el corpus con el objetivo de que se le pudiese extraer todo el valor posible en la parte de modelado. Eliminamos los signos de puntuación, caracteres no alfabéticos y stopwords. y logramos conseguir un corpus que podía ser eficientemente explotado por los modelos posteriores.

- Modelado del corpus:
En esta etapa se trabajaron dos modelos de Machine Learning: Regresión Logística y Clasificación Binaria con Bosques Aleatorios. Se calcularon los mejores hiperparámetros para cada uno y pudimos obtener resultados bastante decentes, todo sin caer en overfitting.

- Reporte de métricas:
Habiendo visto las métricas calculadas para los modelos, podemos extraer conclusiones interesantes. Algunas métricas como el Accuracy o Precision presentaron resultados ligeramente mejores para el modelo de Regresión Logística. Mientras que el Recall presentó resultados ligeramente mejores para el modelo de Clasificación Binaria con Bosques Aleatorios. Esto quiere decir que la selección del mejor modelo dependerá del problema que se este tratando en el momento... En este momento, como no tenemos un objetivo específico, nos dedicaremos a decidir como mejor modelo la Regresión Logística, pues tiene un Accuracy, Precision y F1 Score ligeramente más altos.
